removing unused include, avoiding overflow in flow state refs #1728
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@18217 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/vehicle/SUMOVehicleParameter.h
@@ -437,9 +437,9 @@ public:
     /// @{
 
     /// @brief The number of times the vehicle shall be repeatedly inserted
-    int repetitionNumber;
+    int_fast64_t repetitionNumber;
     /// @brief The number of times the vehicle was already inserted
-    int repetitionsDone;
+    int_fast64_t repetitionsDone;
     /// @brief The time offset between vehicle reinsertions
 #ifdef HAVE_SUBSECOND_TIMESTEPS
     SUMOTime repetitionOffset;
